requirements.txt
accelerate==1.8.1
acvtool==2.3.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.15.2
androguard==4.1.3
apkInspector==1.3.3
Appium-Python-Client==5.0.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
banal==1.0.6
bitsandbytes==0.46.1
certifi==2025.1.31
cffi==1.17.1
Chameleon==4.5.4
charset-normalizer==3.4.1
click==8.1.8
code2flow==2.5.1
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dataset==1.6.2
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
frida==16.7.13
frozenlist==1.7.0
fsspec==2025.3.2
greenlet==3.2.0
h11==0.14.0
huggingface-hub==0.30.2
idna==3.10
ipython==9.1.0
ipython_pygments_lexers==1.1.1
javaobj-py3==0.4.4
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.3.2
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.5.0
mutf8==1.0.6
networkx==3.4.2
numpy==2.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pyaxmlparser==0.3.31
pycparser==2.22
pydot==3.0.4
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.16.0
selenium==4.31.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0+cu126
torch-geometric==2.6.0
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websocket-client==1.8.0
win32_setctime==1.2.0
wsproto==1.2.0
yarl==1.20.1
